Combine DriverPoints

+++ app/src/main/java/com/example/zachcheu/carboncredit/Var.java
@@ -6,13 +6,14 @@ import android.app.Activity;
  * Created by ZachCheu on 10/11/16.
  */
 public class Var extends Activity {
+    //test
     public static int RefreshRate = 0;
     public static boolean Sound = true;
     public static boolean Noti = true;
     public static int Trans = 0;
     public static int[] carsMph = new int[] {30, 40, 50, 60};
     public static int avgMph = 20;
-    public static int idleBase = 4;
+    public static float idleBase = 4;
     public static int localDown = 20;
     public static int localUp  = 40;
     public static int highDown = 50;
